# api/openapi.yaml
openapi: 3.1.0
info:
  title: HeatCheck Public API
  version: 1.0.0
  description: |
    Production API for HeatCheck challenges, proof clips, OpenAI-assisted
    verification, paid participation, voting, moderation, account lifecycle,
    and copyright workflows. All timestamps use RFC 3339 UTC values.
servers:
  - url: https://heatcheck.dogi.watch
    description: Production
  - url: http://localhost:8080
    description: Local development
...
security: []
paths:
...
  /v1/me/subscription:
    get:
      tags: [Billing]
      operationId: getSubscription
      summary: Get subscription status and submission usage
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Server-authoritative subscription and UTC allowance windows.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionOverviewEnvelope"
        "401":
          $ref: "#/components/responses/Error"
  /v1/me/subscription/sync:
    post:
      tags: [Billing]
      operationId: syncSubscription
      summary: Reconcile subscription state after purchase or restore
      description: |
        Fetches current RevenueCat customer information using the authenticated
        HeatCheck user ID as the RevenueCat App User ID.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Reconciled subscription and usage state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionOverviewEnvelope"
        "401":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
        "502":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error:
      description: Structured API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
...
    Subscription:
      type: object
      required:
        - tier
        - entitlement_id
        - status
        - active
        - will_renew
      properties:
        tier:
          type: string
          enum: [none, pro]
        provider:
          type: string
          enum: [revenuecat]
        entitlement_id:
          type: string
        product_id:
          type: string
        store:
          type: string
        environment:
          type: string
          enum: [production, sandbox, unknown]
        status:
          type: string
          enum: [inactive, active, canceled, billing_issue, expired]
        active:
          type: boolean
        will_renew:
          type: boolean
        current_period_start:
          type: [string, "null"]
          format: date-time
        current_period_end:
          type: [string, "null"]
          format: date-time
        management_url:
          type: string
          format: uri
        updated_at:
          type: string
          format: date-time
    SubscriptionUsage:
      type: object
      required:
        - daily_limit
        - daily_used
        - daily_remaining
        - daily_resets_at
        - monthly_limit
        - monthly_used
        - monthly_remaining
        - monthly_resets_at
      properties:
        daily_limit:
          type: integer
        daily_used:
          type: integer
        daily_remaining:
          type: integer
        daily_resets_at:
          type: string
          format: date-time
        monthly_limit:
          type: integer
        monthly_used:
          type: integer
        monthly_remaining:
          type: integer
        monthly_resets_at:
          type: string
          format: date-time
    SubscriptionOverviewEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [app_user_id, subscription, usage]
          properties:
            app_user_id:
              type: string
              format: uuid
              description: Exact App User ID the Flutter RevenueCat SDK must use.
            subscription:
              $ref: "#/components/schemas/Subscription"
            usage:
              $ref: "#/components/schemas/SubscriptionUsage"
...